Revert "Fix RC1004 and API-hash integration in BlindEdr; improve stability and build"

+++ BlindEdr/Common.c
@@ -1,7 +1,6 @@
 #include "Common.h"
 #include "Structs.h"
 #include "FunctionPointers.h"
-#include "RemoveCallBacks.h"
 
 #include <stdio.h>
 
@@ -78,8 +77,8 @@ BOOL saveMemoryFile() {
         }
 
         // Write patch data
-        if (!WriteFile(hFile, p->pData, (DWORD)p->szData, &nNumberOfBytesToWrite, NULL) ||
-            nNumberOfBytesToWrite != (DWORD)p->szData) {
+        if (!WriteFile(hFile, p->pData, p->szData, &nNumberOfBytesToWrite, NULL) ||
+            nNumberOfBytesToWrite != p->szData) {
             CloseHandle(hFile);
             return FALSE;
         }
@@ -143,8 +142,8 @@ BOOL restoreBlindness() {
         }
 
         // Read patch data
-        if (!ReadFile(hFile, patchData, (DWORD)pt.szData, &bytesRead, NULL) ||
-            bytesRead != (DWORD)pt.szData) {
+        if (!ReadFile(hFile, patchData, pt.szData, &bytesRead, NULL) ||
+            bytesRead != pt.szData) {
             success = FALSE;
             break;
         }
